Move to src and add board class

+++ src/ship.cpp
@@ -1,10 +1,10 @@
-//
-// Created by Jason on 2/17/2019.
-//
+// ship.cpp
+// Simeon Ng
+// Jason Herning
+// Updated: 2/22/19
+// Ship Class Source File
 
 #include "ship.h"
-
-
 
 Ship::Ship(size_t position, size_t size):_position(position),_size(size)
 {
@@ -21,14 +21,13 @@ std::size_t Ship::getSize()
     return _size;
 }
 
-void Ship::hit(int position)
+void Ship::hit(size_t position)
 {
     _hits[position] = true;
 }
 
 
-bool Ship::getHits(int position)
+bool Ship::getHits(size_t position)
 {
     return _hits[position];
-
 }
